requirements.txt
requests
pillow
pyzbar
json
pandas
numpy
collections